backend/pyproject.toml
[project]
name = "ytplayer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "flask-cors>=5.0.0",
    "flask>=3.0.3",
    "pymongo>=4.10.1",
    "python-dotenv>=1.0.1",
    "typing-extensions>=4.12.2",
    "yt-dlp>=2024.10.7",
]
